feat(news): Fixing decorator - front-3139
